requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.12
langchain-openai>=0.1.22